move resttemplate config to new package web
