+++ src/main/groovy/rocks/metaldetector/butler/config/web/CustomClientErrorHandler.groovy
@@ -1,4 +1,4 @@
-package rocks.metaldetector.butler.config.resttemplate
+package rocks.metaldetector.butler.config.web
 
 import groovy.util.logging.Slf4j
 import org.springframework.http.HttpMethod
